ai-engine/requirements.txt
fastapi
uvicorn
pydantic
langgraph
langchain
pinecone
sentence-transformers
PyMuPDF
requests
python-dotenv
langchain-groq
langchain-text-splitters
langchain-community